refactor: move FileError to protocols errors
