+++ src/protocol/vm/mod.rs
@@ -2,7 +2,6 @@ mod adapter_contract;
 pub mod constants;
 pub mod engine;
 pub(crate) mod erc20_overwrite_factory;
-pub mod errors;
 mod models;
 pub mod state;
 pub mod state_builder;
